requirements.txt
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.11.0
flask_api==3.0.post1
python-dotenv==1.0.0
Flask-Limiter==3.3.1